Move NumberPicker into separate folder and convert it into a functional component
